can now delete moods aswell + moved moodColor function into util folder
