+++ package.json
@@ -19,7 +19,9 @@
     "react-native-safe-area-context": "4.10.5",
     "react-native-screens": "3.31.1",
     "expo-router": "~3.5.23",
-    "@react-native-async-storage/async-storage": "1.23.1"
+    "@react-native-async-storage/async-storage": "1.23.1",
+    "react-native-web": "~0.19.10",
+    "react-dom": "18.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
